REFACTOR(Auth): Reorganize DTOs and depure AuthService/Controller
- Repositioned `UserRegisterStepOneDTO` and `UserRegisterStepTwoDTO` to **dto.auth** package for better organization within the project.

- Depured `AuthController` and `AuthService` to improve code quality and consistency.
