+++ src/main/java/com/copay/app/dto/auth/UserRegisterStepOneDTO.java
@@ -1,4 +1,4 @@
-package com.copay.app.dto;
+package com.copay.app.dto.auth;
 
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.Email;
